% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_more_columns}
\alias{add_more_columns}
\title{Add more columns}
\usage{
add_more_columns(res)
}
\arguments{
\item{res}{A \code{enrichResult} object or a \code{gseaResult} object}
}
\description{
Add more columns
}
\details{
It adds the following columns in the enrichment table:
\itemize{
\item n_hits
\item n_genes
\item gs_size
\item n_total
\item log2_fold_enrichment
\item z_score
}
}
