% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert_to_entrez_id}
\alias{convert_to_entrez_id}
\title{Convert to EntreZ IDs}
\usage{
convert_to_entrez_id(x, org_db = "org.Hs.eg.db")
}
\arguments{
\item{x}{A vector of gene IDs, a named numeric vector where gene IDs are names, or a matrix with gene IDs as row names}

\item{org_db}{The name of the \code{OrgDb} object.}
}
\description{
Convert to EntreZ IDs
}
