% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea.R
\name{fgsea_wrapper}
\alias{fgsea_wrapper}
\alias{fgsea_go}
\alias{fgsea_kegg}
\alias{fgsea_msigdb}
\alias{fgsea_reactome}
\alias{fgsea_keywords}
\alias{fgsea_phenotype}
\alias{fgsea_disease}
\title{GSEA using fgsea}
\usage{
fgsea_wrapper(s, gs, min_size = 5, max_size = 2000, ...)

fgsea_go(s, org_db = org.Hs.eg.db::org.Hs.eg.db, ontology = "BP", ...)

fgsea_kegg(s, organism = "hsa", db = "pathway", ...)

fgsea_msigdb(s, collection = "h.all", version = "2024.1.Hs", ...)

fgsea_reactome(s, organism = "HSA", ...)

fgsea_keywords(s, organism = "human", ...)

fgsea_phenotype(s, organism = "human", ...)

fgsea_disease(s, organism = "human", ...)
}
\arguments{
\item{s}{A numeric vector of gene scores. It should be named with gene IDs.}

\item{gs}{A list of gene sets. In \code{fgsea_wrapper()}, genes should have the same ID types as \code{s}.}

\item{min_size}{Minimal size of gene sets for analysis.}

\item{max_size}{Maximal size of gene sets for analysis.}

\item{...}{Other argument passed to \code{fgsea_wrapper()} or further to \code{\link[fgsea:fgsea]{fgsea::fgsea()}}.}

\item{org_db}{An \code{OrgDb} object for the organism. It can be from \strong{org.*.db} packages or downloaded by the \strong{AnnotationHub} package.}

\item{ontology}{Namespace of GO. Value should be one of "BP", "CC" or "MF".}

\item{organism}{See \strong{Details}.}

\item{db}{A KEGG database. The value can be one of "pathway", "module", "ko", "network", "disease" and "drug".}

\item{collection}{Collection of the MSigDB gene sets. All possible values can be found via \code{\link[=list_msigdb_versions]{list_msigdb_versions()}}.}

\item{version}{Version of the MSigDB database. All possible values can be found via \code{\link[=list_msigdb_collections]{list_msigdb_collections()}}.}
}
\description{
GSEA using fgsea
}
\details{
Except \code{fgsea_wrapper()}, gene IDs in \code{s} in all \verb{fgsea_*()} functions must be EntreZ IDs.

The value should be set differently for specific \verb{fgsea_*()} functions.
\itemize{
\item for \code{fgsea_kegg()}, the value should be a KEGG organism code, such as "hsa" or "mmu".
\item for \code{fgsea_reactome()}, the value should a prefix of the Reactome pathway ID that represents the organism. E.g. "HSA" for human.
\item for \code{fgsea_keywords()}, the value can be a organism name, e.g. "human", the latin name or the taxon ID. Please check \code{\link[UniProtKeywords:load_keyword_genesets]{UniProtKeywords::load_keyword_genesets()}}.
\item for \code{fgsea_phenotype()} and \code{fgsea_disease()}, the value can only be one of "human", "mouse" and "rat".
}

All valid values for \code{fgsea_reactome()} are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{c("BTA", "CEL", "CFA", "DRE", "DDI", "DME", "GGA", "HSA", "MMU", 
  "MTU", "PFA", "RNO", "SCE", "SPO", "SSC", "XTR")
}\if{html}{\out{</div>}}
}
\examples{
data(p53_dataset)
s = p53_dataset$s2n
gs = p53_dataset$gs

fgsea_wrapper(s, gs) |> head()

s2 = convert_to_entrez(s)

fgsea_go(s2) |> head()
fgsea_kegg(s2) |> head()
fgsea_msigdb(s2) |> head()
}
