% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgdb.R
\name{get_GO_gene_sets_from_orgdb}
\alias{get_GO_gene_sets_from_orgdb}
\title{Get GO gene sets from OrgDb object}
\usage{
get_GO_gene_sets_from_orgdb(
  org_db,
  ontology = "BP",
  gene_id_type = c("ENTREZID", "SYMBOL", "ENSEMBL"),
  as_table = FALSE
)
}
\arguments{
\item{org_db}{An \code{OrgDb} object.}

\item{ontology}{"BP", "CC" or "MF".}

\item{gene_id_type}{Which gene ID type to use.}

\item{as_table}{Whether to return a list or a two-column data frame?}
}
\description{
Get GO gene sets from OrgDb object
}
