% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\name{gs_map_to_orthologues}
\alias{gs_map_to_orthologues}
\title{Generate gene sets by mapping to orthologues}
\usage{
gs_map_to_orthologues(gs, from, to)
}
\arguments{
\item{gs}{A list of gene sets. Genes should be in the EntreZ ID type.}

\item{from}{Latin name of the "from" organism.}

\item{to}{Latin name of the "to" organism.}
}
\description{
Generate gene sets by mapping to orthologues
}
\details{
Gene IDs in \code{gs} must be EntreZ IDs.

Gene sets are removed if none of the genes can ba mapped.
}
\examples{
gs_human = load_msigdb_genesets()$gs
gs_dolphin = gs_map_to_orthologues(gs_human, from = "Homo.sapiens", 
    to = "Tursiops.truncatus")
}
