% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gsea}
\alias{gsea}
\alias{gsea_gene_perm}
\title{GSEA analysis}
\usage{
gsea(expr, condition, cmp, geneset, plot = FALSE, power = 1)

gsea_gene_perm(s, geneset, perm = FALSE, power = 1)
}
\arguments{
\item{expr}{The complete expression matrix}

\item{condition}{The condition labels of samples}

\item{cmp}{A vector of two, \code{cmp[1] - cmp[2] > 0} means up-regulation}

\item{geneset}{A vector of genes}

\item{plot}{Whether to make the GSEA plot?}

\item{power}{Power added to the weight.}

\item{s}{A pre-sorted gene-level scores.}

\item{perm}{Whether to perform gene-permutation?}
}
\description{
GSEA analysis
}
