% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework.R
\name{gene_level_log2fc}
\alias{gene_level_log2fc}
\alias{gene_level_s2n}
\alias{gene_level_tvalue}
\alias{gene_level_sam}
\alias{gene_level_log_p}
\alias{set_level_mean}
\alias{set_level_median}
\alias{set_level_maxmean}
\alias{set_level_wilcox}
\alias{set_chi_square}
\alias{set_level_ks}
\title{GSEA framework}
\usage{
gene_level_log2fc(mat, condition)

gene_level_s2n(mat, condition)

gene_level_tvalue(mat, condition)

gene_level_sam(mat, condition)

gene_level_log_p(mat, condition)

set_level_mean(s, l)

set_level_median(s, l)

set_level_maxmean(s, l)

set_level_wilcox(s, l)

set_chi_square(s, l)

set_level_ks(s, l)
}
\arguments{
\item{mat}{The input matrix.}

\item{condition}{The condition labels of samples. The value should be a factor where \code{level[1]} corresponds to the treatment group and \code{level[2]} corresponds to the control group.}

\item{s}{A vector of gene scores.}

\item{l}{A logical vector representing whether genes in \code{s} are in a specific gene set.}
}
\description{
GSEA framework
}
\details{
There are the following gene-level statistics
\itemize{
\item log2fc
\item signal to noise ratio
\item t-value
\item SAM t-value
\item -log(p)
}

There are the following gene set level statistics:
\itemize{
\item mean
\item median
\item maxmean
\item wilcox ranksum
\item ks. Since we only support one-side test, the ES is defined as \code{max(abs(f2 - f1))}.
\item chi-square. Only works when gene-level scores are binary.
}
}
