% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_id.R
\name{map_to_entrez_id}
\alias{map_to_entrez_id}
\alias{guess_id_type}
\alias{convert_to_entrez_id}
\title{ID mapping}
\usage{
map_to_entrez_id(from, org_db = org.Hs.eg.db::org.Hs.eg.db)

guess_id_type(id, org_db = org.Hs.eg.db::org.Hs.eg.db, verbose = TRUE)

convert_to_entrez_id(x, org_db = org.Hs.eg.db::org.Hs.eg.db)
}
\arguments{
\item{from}{The "keytype" of the "from" ID type.}

\item{org_db}{An \code{OrgDb} object.}

\item{id}{A vector of gene IDs}

\item{verbose}{Whether to print messages?}

\item{x}{It supports three formats: 1. A vector of gene IDs, 2. A numeric vector with gene IDs as names, 3. A numeric matrix with gene IDs as row names.}
}
\description{
ID mapping
}
\details{
\code{map_to_entrez_id()} returns a gene ID mapping from the "from" type to EntreZ IDs, only for protein-coding genes.

\code{guess_id_type()} automatically tests the following ID types: ENTREZID, ENSEMBL (genes), REFSEQ and SYMBOL.

For \code{convert_to_entrez_id()}, if mapping is applied on a character gene ID vector, IDs that cannot be mapped are removed, and for duplicated mappings only
the first one is picked. If mapping is applied on a numeric vector or a matrix, and if there are multiple mappings, the mean is taken as the final value.
}
