% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{interactive_local}
\alias{interactive_local}
\title{A REPL in a local environment}
\usage{
interactive_local(parent = parent.frame())
}
\arguments{
\item{parent}{Parent environment.}
}
\description{
A REPL in a local environment
}
