% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{jaccard}
\alias{jaccard}
\alias{genesets_similarity_heatmap}
\title{Jaccard coefficent}
\usage{
jaccard(lt, weight = NULL)

genesets_similarity_heatmap(
  lt,
  weight = NULL,
  min_pct = 0.5,
  max_k = 5,
  resolution = 1
)
}
\arguments{
\item{lt}{A list of gene sets.}

\item{weight}{Weight of genes. The value should be a named vector and it should cover all genes in \code{lt}.}

\item{min_pct}{Minimal recurreny of genes in a geneset cluster.}

\item{max_k}{Maximal number of recurrent genes to use on the plot.}

\item{resolution}{For controlling the graph clustering method.}
}
\description{
Jaccard coefficent
}
