% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\name{load_go_genesets}
\alias{load_go_genesets}
\alias{load_kegg_genesets}
\alias{load_msigdb_genesets}
\alias{load_reactome_genesets}
\alias{load_simona_genesets}
\title{Load various gene sets}
\usage{
load_go_genesets(org_db, ontology = "BP")

load_kegg_genesets(organism, db = "pathway")

load_msigdb_genesets(collection = "h.all", version = "2024.1.Hs")

load_reactome_genesets(organism)

load_simona_genesets(fun, ontology, organism = "human")
}
\arguments{
\item{org_db}{An \code{OrgDb} object.}

\item{ontology}{Just a tag, please use \code{"hp"} or \code{"rdo"}.}

\item{organism}{See \strong{Details}.}

\item{db}{A KEGG database. The value can be one of "pathway", "module", "ko", "network", "disease" and "drug".}

\item{collection}{Collection of the MSigDB gene sets. All possible values can be found via \code{\link[=list_msigdb_versions]{list_msigdb_versions()}}.}

\item{version}{Version of the MSigDB database. All possible values can be found via \code{\link[=list_msigdb_collections]{list_msigdb_collections()}}.}

\item{fun}{Use \code{\link[simona:ontology]{simona::ontology_hp}} or \code{\link[simona:ontology]{simona::ontology_rdo}}.}
}
\description{
Load various gene sets
}
\details{
Genes are all in the EntreZ ID type. All the gene sets are saved in cache.

The value should be set differently for specific \verb{fgsea_*()} functions.
\itemize{
\item for \code{load_kegg_genesets()}, the value should be a KEGG organism code, such as "hsa" or "mmu".
\item for \code{load_reactome_genesets()}, the value should a prefix of the Reactome pathway ID that represents the organism. E.g. "HSA" for human.
\item for \code{load_simona_genesets()}, the value can only be one of "human", "mouse" and "rat".
}

All valid values for \code{load_reactome_genesets()} are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{c("BTA", "CEL", "CFA", "DRE", "DDI", "DME", "GGA", "HSA", "MMU", 
  "MTU", "PFA", "RNO", "SCE", "SPO", "SSC", "XTR")
}\if{html}{\out{</div>}}

The human phenotype gene sets and disease gene sets are supported by the \strong{simona} package.
}
