% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea.R
\name{gsea_sample_perm}
\alias{gsea_sample_perm}
\alias{gsea_gene_perm}
\title{Native implementation of GSEA}
\usage{
gsea_sample_perm(
  expr,
  condition,
  gs,
  direction = c("std", "pos", "neg"),
  power = 1,
  min_size = 5,
  max_size = 1000,
  nperm = 1000,
  verbose = TRUE
)

gsea_gene_perm(
  s,
  gs,
  direction = c("std", "pos", "neg"),
  power = 1,
  min_size = 5,
  max_size = 1000,
  nperm = 1000,
  verbose = TRUE
)
}
\arguments{
\item{expr}{The complete expression matrix.}

\item{condition}{The condition labels of samples. The value should be a factor where \code{level[1]} corresponds to the treatment group and \code{level[2]} corresponds to the control group.}

\item{gs}{A list of gene sets. Genes should have the same ID types as \code{expr}.}

\item{direction}{Whether the test is one-sided or two-sided.}

\item{power}{Power added to the absolute value of weight.}

\item{min_size}{Minimal size of gene sets for analysis.}

\item{max_size}{Maximal size of gene sets for analysis.}

\item{nperm}{Number of permutations.}

\item{verbose}{Whether to print messages?}

\item{s}{A numeric vector of gene scores.}
}
\description{
Native implementation of GSEA
}
\details{
The two functions are only for the purpose of studying.
}
