% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgdb.R
\name{get_GO_gene_sets_from_orgdb}
\alias{get_GO_gene_sets_from_orgdb}
\alias{random_genes}
\alias{protein_coding_genes}
\title{Get genes and GO terms from OrgDb oject}
\usage{
get_GO_gene_sets_from_orgdb(
  org_db,
  ontology = "BP",
  gene_id_type = c("ENTREZID", "SYMBOL", "ENSEMBL"),
  as_table = FALSE
)

random_genes(
  org_db,
  n = 1000,
  gene_id_type = c("ENTREZID", "SYMBOL", "ENSEMBL")
)

protein_coding_genes(org_db, gene_id_type = c("ENTREZID", "SYMBOL", "ENSEMBL"))
}
\arguments{
\item{org_db}{An \code{OrgDb} object.}

\item{ontology}{"BP", "CC" or "MF".}

\item{gene_id_type}{Which gene ID type to use?}

\item{as_table}{Whether to return a list or a two-column data frame?}

\item{n}{Number of random genes.}
}
\description{
Get genes and GO terms from OrgDb oject
}
\examples{
require(org.Hs.eg.db)
gs = get_GO_gene_sets_from_orgdb(org.Hs.eg.db)
gs[1:2]
random_genes(org.Hs.eg.db) |> head()
protein_coding_genes(org.Hs.eg.db) |> head()
}
