% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgdb.R
\name{random_genes}
\alias{random_genes}
\title{Random genes}
\usage{
random_genes(org_db, n = 1000, keytype = "ENTREZID")
}
\arguments{
\item{org_db}{An \code{OrgDb} object.}

\item{n}{Number of random genes.}

\item{keytype}{Keytype of the genes in the \code{OrgDb} database.}
}
\description{
Random genes
}
