% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seqlengths_as_GRanges}
\alias{seqlengths_as_GRanges}
\title{Create a GRanges object from seqlengths}
\usage{
seqlengths_as_GRanges(x, seqnames = NULL)
}
\arguments{
\item{x}{Any object that supports the \code{seqlengths()} method.}

\item{seqnames}{A vector seqnames.}
}
\description{
Create a GRanges object from seqlengths
}
\details{
It basically generates a \code{GRanges} treating whole chromosomes as genomic intervals.
}
\examples{
require(TxDb.Hsapiens.UCSC.hg19.knownGene)
seqlengths_as_GRanges(TxDb.Hsapiens.UCSC.hg19.knownGene)
seqlengths_as_GRanges(TxDb.Hsapiens.UCSC.hg19.knownGene, 
    paste0("chr", c(1:22, "X", "Y")))
}
