% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{term_similarity}
\alias{term_similarity}
\title{Similarity between terms based on overlap}
\usage{
term_similarity(
  gl,
  method = c("kappa", "jaccard", "dice", "overlap"),
  all = NULL,
  remove_negative = TRUE
)
}
\arguments{
\item{gl}{A list of items}

\item{method}{The similarity measurement.}

\item{all}{The universe set.}

\item{remove_negative}{If the value is negative, reset to zero}
}
\value{
A symmetric matrix.
}
\description{
Similarity between terms based on overlap
}
