% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\name{gs_map_to_orthologues}
\alias{gs_map_to_orthologues}
\title{Generate gene sets by mapping to orthologues}
\usage{
gs_map_to_orthologues(gs, from, to)
}
\arguments{
\item{gs}{A list of gene sets. Genes should be in the EntreZ ID type.}

\item{from}{Latin name of the "from" organism.}

\item{to}{Latin name of the "to" organism.}
}
\description{
Generate gene sets by mapping to orthologues
}
