% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_to_dataframe}
\alias{list_to_dataframe}
\alias{dataframe_to_list}
\title{Convert gene sets between list and a data frame}
\usage{
list_to_dataframe(list)

dataframe_to_list(df)
}
\arguments{
\item{list}{A list of genes.}

\item{df}{A two-column data frame}
}
\description{
Convert gene sets between list and a data frame
}
\details{
In \code{gs_dataframe_to_list()}, which column contains genes and which column contains gene sets
are automatically checked by the number of genes and gene sets. Basically number of genes
should be larger than the number of gene sets.
}
