% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_msigdb_versions}
\alias{list_msigdb_versions}
\alias{list_msigdb_collections}
\alias{get_msigdb}
\title{Retrieve gene sets from MSigDB}
\usage{
list_msigdb_versions()

list_msigdb_collections(version = NULL)

get_msigdb(
  version = choose_msigdb_version(),
  collection = choose_msigdb_collection(version),
  gene_id_type = c("entrez", "symbols"),
  as_table = FALSE
)
}
\arguments{
\item{version}{The MSigDB version. The value can be obtained by \code{list_msigdb_versions()}.
If this argument is not specified, a menu will be printed to let users to select.}

\item{collection}{The gene set collection. The values can be obtained by \code{list_msigdb_collections()}.}

\item{gene_id_type}{One of "entrez" and "symbols".}

\item{as_table}{Whether to return a list or a table.}
}
\description{
Retrieve gene sets from MSigDB
}
