% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ora.R
\name{ora}
\alias{ora}
\alias{ora_go}
\alias{ora_kegg}
\alias{ora_msigdb}
\alias{ora_reactome}
\alias{ora_keywords}
\alias{ora_phenotype}
\alias{ora_disease}
\title{Over-representation analysis}
\usage{
ora(genes, gs, universe = NULL, min_hits = 3, min_size = 5, max_size = 2000)

ora_go(genes, org_db = org.Hs.eg.db::org.Hs.eg.db, ontology = "BP", ...)

ora_kegg(genes, organism = "hsa", db = "pathway", ...)

ora_msigdb(genes, collection = "h.all", version = "2024.1.Hs", ...)

ora_reactome(genes, organism = "HSA", ...)

ora_keywords(genes, organism = "human", ...)

ora_phenotype(genes, organism = "human", ...)

ora_disease(genes, organism = "human", ...)
}
\arguments{
\item{genes}{A vector of genes.}

\item{gs}{A list of gene sets. Genes should have the smae ID type as in \code{genes}.}

\item{universe}{A vector of background genes.}

\item{min_hits}{Minimal number of overlapping genes in \code{genes} and gene sets.}

\item{min_size}{Minimal number of genes in gene sets.}

\item{max_size}{Maximal number of genes in gene sets.}

\item{org_db}{An \code{OrgDb} object for the organism. It can be from \strong{org.*.db} packages or downloaded by the \strong{AnnotationHub} package.}

\item{ontology}{Namespace of GO. Value should be one of "BP", "CC" or "MF".}

\item{...}{Passed to \code{ora()}.}

\item{organism}{See \strong{Details}.}

\item{db}{A KEGG database. The value can be one of "pathway", "module", "ko", "network", "disease" and "drug".}

\item{collection}{Collection of the MSigDB gene sets. All possible values can be found via \code{\link[=list_msigdb_versions]{list_msigdb_versions()}}.}

\item{version}{Version of the MSigDB database. All possible values can be found via \code{\link[=list_msigdb_collections]{list_msigdb_collections()}}.}
}
\description{
Over-representation analysis
}
\details{
Except \code{ora()}, gene IDs in \code{s} in all \verb{ora_*()} functions must be EntreZ IDs.

The value should be set differently for specific \verb{ora_*()} functions.
\itemize{
\item for \code{ora_kegg()}, the value should be a KEGG organism code, such as "hsa" or "mmu".
\item for \code{ora_reactome()}, the value should a prefix of the Reactome pathway ID that represents the organism. E.g. "HSA" for human.
\item for \code{ora_keywords()}, the value can be a organism name, e.g. "human", the latin name or the taxon ID.
\item for \code{ora_phenotype()} and \code{fgsea_disease()}, the value can only be one of "human", "mouse" and "rat".
}

All valid values for \code{fgsea_reactome()} are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{c("BTA", "CEL", "CFA", "DRE", "DDI", "DME", "GGA", "HSA", "MMU", 
  "MTU", "PFA", "RNO", "SCE", "SPO", "SSC", "XTR")
}\if{html}{\out{</div>}}
}
\examples{
data(p53_dataset)
s = p53_dataset$s2n
gs = p53_dataset$gs
diff = names(s)[abs(s) > 0.3]

ora(diff, gs) |> head()

diff2 = convert_to_entrez(diff)
ora_go(diff2) |> head()
ora_msigdb(diff2) |> head()
}
