% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ora}
\alias{ora}
\title{Over-representation analysis}
\usage{
ora(genes, gene_sets, universe = NULL, min_hits = 3)
}
\arguments{
\item{genes}{A vector of genes.}

\item{gene_sets}{A list of vectors.}

\item{universe}{A vector of background genes.}

\item{min_hits}{Minimal number of genes in \code{genes} in a gene set.}
}
\description{
Over-representation analysis
}
