% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_gsea.R
\name{ss_gsea}
\alias{ss_gsea}
\title{Single-sample GSEA transformation}
\usage{
ss_gsea(mat, gs, scale = TRUE, set_level = set_level_mean)
}
\arguments{
\item{mat}{A numeric matrix.}

\item{gs}{A list of gene sets. Genes should have the same ID type as in \code{mat}.}

\item{scale}{Whether to apply z-score scaling on rows?}

\item{set_level}{A function to calculate the geneset-level score. This function should accept two arguments: a vector of scores for all genes, and a corresponding logical vector repressenting which genes are in the current gene set.}
}
\description{
Single-sample GSEA transformation
}
