% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{retrieve_all_kegg_graphs}
\alias{retrieve_all_kegg_graphs}
\alias{ora_kegg_topology}
\alias{gsea_kegg_topology}
\title{Centrality-based KEGG enrichment analysis}
\usage{
retrieve_all_kegg_graphs(organism = "hsa")

ora_kegg_topology(
  genes,
  universe = NULL,
  centrality = igraph::degree,
  organism = "hsa",
  pl = NULL,
  perm = 1000,
  min_hits = 3,
  min_size = 5,
  max_size = 2500,
  verbose = TRUE
)

gsea_kegg_topology(
  s,
  centrality = igraph::degree,
  organism = "hsa",
  pl = NULL,
  min_size = 5,
  max_size = 2500,
  verbose = TRUE
)
}
\arguments{
\item{organism}{KEGG organism code.}

\item{genes}{A vector of genes.}

\item{universe}{A vector of universe genes. If it is not specified, the total genes in the KEGG pathways are used.}

\item{centrality}{Centrality method. The value should be a function which accepts an \code{igraph} object and returns a vector of centrality values.}

\item{pl}{A list of KEGG pathways as \code{igraph} objects. If the organism is one of "hsa", "mmu" or "rno", the corresponding \code{pl} is already
generated and will be loaded automatically. For other organisms, use \code{retrieve_all_kegg_graphs()} to generate one.}

\item{perm}{Number of permutations.}

\item{min_hits}{Minimal number of the overlapping genes in \code{genes} and pathways.}

\item{min_size}{Minimal number of genes in pathways.}

\item{max_size}{Maximal number of genes in pathways.}

\item{verbose}{Whether to print messages?}

\item{s}{A numeric vector of gene scores.}
}
\description{
Centrality-based KEGG enrichment analysis
}
\details{
The following lists several useful centrality measures, written as functions. They can be assigned to the \code{centrality} argument.
\itemize{
\item in-degree: \code{function(g) igraph::degree(g, mode = "in")}
\item out-degree: \code{function(g) igraph::degree(g, mode = "out")}
\item betweenness: \code{igraph::betweenness}
\item page rank: \code{igraph::page_rank}
\item in-reach: \code{function(g) CePa::reach(g, mode = "in")}
\item out-reach: \code{function(g) CePa::reach(g, mode = "out")}
}

And many more in the \strong{igraph} and \strong{CePa} packages.

\code{ora_kegg_topology()} is the ORA-extension.

\code{gsea_kegg_topology()} is the GSEA extension. The geneset-level score is calculated as \code{mean(abs(s*w))}.
}
