% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap_text}
\alias{wrap_text}
\title{Wrap long text into several lines}
\usage{
wrap_text(x, width = 60)
}
\arguments{
\item{x}{A vector of sentences.}

\item{width}{Maximal number of chararacters per line.}
}
\description{
Wrap long text into several lines
}
